requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.0
pyyaml>=6.0
python-json-logger>=2.0.0
structlog>=24.0.0
python-dotenv>=1.0.0
